+++ plugins/ikfastsolvers/ikbase.h
@@ -521,6 +521,18 @@ private:
                 IKReal eerot[9] = {p.second, 0, 0, 0, 0, 0, 0, 0, 0};
                 return _pfnik(eetrans, eerot, vfree.size()>0 ? &vfree[0] : NULL, vsolutions);
             }
+            case IKP_TranslationYAxisAngle4D: {
+                std::pair<Vector,dReal> p = param.GetTranslationYAxisAngle4D();
+                IKReal eetrans[3] = {p.first.x, p.first.y,p.first.z};
+                IKReal eerot[9] = {p.second, 0, 0, 0, 0, 0, 0, 0, 0};
+                return _pfnik(eetrans, eerot, vfree.size()>0 ? &vfree[0] : NULL, vsolutions);
+            }
+            case IKP_TranslationZAxisAngle4D: {
+                std::pair<Vector,dReal> p = param.GetTranslationZAxisAngle4D();
+                IKReal eetrans[3] = {p.first.x, p.first.y,p.first.z};
+                IKReal eerot[9] = {p.second, 0, 0, 0, 0, 0, 0, 0, 0};
+                return _pfnik(eetrans, eerot, vfree.size()>0 ? &vfree[0] : NULL, vsolutions);
+            }
             default:
                 BOOST_ASSERT(0);
                 break;